refactor: moved out types
