+++ cognite/neat/rules/models/entities/_base.py
@@ -3,7 +3,7 @@ import sys
 from abc import ABC, abstractmethod
 from functools import total_ordering
 from types import UnionType
-from typing import Annotated, Any, ClassVar, Generic, Literal, TypeVar, Union, cast, get_args, get_origin
+from typing import Any, ClassVar, Generic, Literal, TypeVar, Union, cast, get_args, get_origin
 
 from cognite.client.data_classes.data_modeling import DirectRelationReference
 from cognite.client.data_classes.data_modeling.ids import (
@@ -14,11 +14,8 @@ from cognite.client.data_classes.data_modeling.ids import (
     ViewId,
 )
 from pydantic import (
-    AnyHttpUrl,
     BaseModel,
-    BeforeValidator,
     Field,
-    PlainSerializer,
     model_serializer,
     model_validator,
 )
@@ -510,80 +507,3 @@ class ReferenceEntity(ClassEntity):
 
     def as_class_entity(self) -> ClassEntity:
         return ClassEntity(prefix=self.prefix, suffix=self.suffix, version=self.version)
-
-
-def _split_str(v: Any) -> list[str]:
-    if isinstance(v, str):
-        return v.replace(", ", ",").split(",")
-    return v
-
-
-def _join_str(v: list[ClassEntity]) -> str | None:
-    return ",".join([entry.id for entry in v]) if v else None
-
-
-def _generate_cdf_resource_list(v: Any) -> list[AssetEntity | RelationshipEntity]:
-    results = []
-    for item in _split_str(v):
-        if isinstance(item, str):
-            if "relationship" in item.lower():
-                results.append(RelationshipEntity.load(item))
-            elif "asset" in item.lower():
-                results.append(AssetEntity.load(item))  # type: ignore
-            else:
-                raise ValueError(f"Unsupported implementation definition: {item}")
-
-        elif isinstance(item, AssetEntity | RelationshipEntity):
-            results.append(item)
-        else:
-            raise ValueError(f"Unsupported implementation definition: {item}")
-
-    return results  # type: ignore
-
-
-ClassEntityList = Annotated[
-    list[ClassEntity],
-    BeforeValidator(_split_str),
-    PlainSerializer(
-        _join_str,
-        return_type=str,
-        when_used="unless-none",
-    ),
-]
-
-
-CdfResourceEntityList = Annotated[
-    list[AssetEntity | RelationshipEntity],
-    BeforeValidator(_generate_cdf_resource_list),
-    PlainSerializer(
-        _join_str,
-        return_type=str,
-        when_used="unless-none",
-    ),
-]
-
-
-ContainerEntityList = Annotated[
-    list[ContainerEntity],
-    BeforeValidator(_split_str),
-    PlainSerializer(
-        _join_str,
-        return_type=str,
-        when_used="unless-none",
-    ),
-]
-
-ViewEntityList = Annotated[
-    list[ViewEntity],
-    BeforeValidator(_split_str),
-    PlainSerializer(
-        _join_str,
-        return_type=str,
-        when_used="unless-none",
-    ),
-]
-
-URLEntity = Annotated[
-    AnyHttpUrl,
-    PlainSerializer(lambda v: str(v), return_type=str, when_used="unless-none"),
-]
